// web/src/api/generated/models/QuietHoursUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface QuietHoursUpdate
 */
export interface QuietHoursUpdate {
    /**
     * 
     * @type {boolean}
     * @memberof QuietHoursUpdate
     */
    enabled: boolean;
    /**
     * 
     * @type {string}
     * @memberof QuietHoursUpdate
     */
    end?: string | null;
    /**
     * 
     * @type {string}
     * @memberof QuietHoursUpdate
     */
    start?: string | null;
}

/**
 * Check if a given object implements the QuietHoursUpdate interface.
 */
export function instanceOfQuietHoursUpdate(value: object): value is QuietHoursUpdate {
    if (!('enabled' in value) || value['enabled'] === undefined) return false;
    return true;
}

export function QuietHoursUpdateFromJSON(json: any): QuietHoursUpdate {
    return QuietHoursUpdateFromJSONTyped(json, false);
}

export function QuietHoursUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): QuietHoursUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'enabled': json['enabled'],
        'end': json['end'] === undefined ? undefined : json['end'] === null ? null : json['end'],
        'start': json['start'] === undefined ? undefined : json['start'] === null ? null : json['start'],
    };
}

export function QuietHoursUpdateToJSON(json: any): QuietHoursUpdate {
    return QuietHoursUpdateToJSONTyped(json, false);
}

export function QuietHoursUpdateToJSONTyped(value?: QuietHoursUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'enabled': value['enabled'],
        'end': value['end'],
        'start': value['start'],
    };
}